refactor: export functions
